feat: create shared utility file
